backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
openai
python-dotenv
textstat
markdownify
fastapi-cors
google-generativeai